+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+- Date scalars based on `DateScalar` now support `\Carbon\CarbonImmutable`
+
 ## v5.21.0
 
 ### Added